refactor: change folder organization
